// src/generated/graphql.ts
export type App = Node & {
  __typename?: 'App';
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars['DateTime'];
  /** Identifies the primary key from the database. */
  databaseId?: Maybe<Scalars['Int']>;
  /** The description of the app. */
  description?: Maybe<Scalars['String']>;
  id: Scalars['ID'];
  /** The hex color code, without the leading '#', for the logo background. */
  logoBackgroundColor: Scalars['String'];
  /** A URL pointing to the app's logo. */
  logoUrl: Scalars['URI'];
  /** The name of the app. */
  name: Scalars['String'];
  /** A slug based on the name of the app for use in URLs. */
  slug: Scalars['String'];
  /** Identifies the date and time when the object was last updated. */
  updatedAt: Scalars['DateTime'];
  /** The URL to the app's homepage. */
  url: Scalars['URI'];
};
// ...
export type Issue = Node &
  Assignable &
  Closable &
  Comment &
  Updatable &
  UpdatableComment &
  Labelable &
  Lockable &
  Reactable &
  RepositoryNode &
  Subscribable &
  UniformResourceLocatable & {
    __typename?: 'Issue';
    /** Reason that the conversation was locked. */
    activeLockReason?: Maybe<LockReason>;
    /** A list of Users assigned to this object. */
    assignees: UserConnection;
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** Identifies the body of the issue. */
    body: Scalars['String'];
    /** The body rendered to HTML. */
    bodyHTML: Scalars['HTML'];
    /** The http path for this issue body */
    bodyResourcePath: Scalars['URI'];
    /** Identifies the body of the issue rendered to text. */
    bodyText: Scalars['String'];
    /** The http URL for this issue body */
    bodyUrl: Scalars['URI'];
    /** `true` if the object is closed (definition of closed may depend on type) */
    closed: Scalars['Boolean'];
    /** Identifies the date and time when the object was closed. */
    closedAt?: Maybe<Scalars['DateTime']>;
    /** A list of comments associated with the Issue. */
    comments: IssueCommentConnection;
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars['DateTime'];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars['Boolean'];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars['Int']>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    /** The hovercard information for this issue */
    hovercard: Hovercard;
    id: Scalars['ID'];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars['Boolean'];
    /** Is this issue read by the viewer */
    isReadByViewer?: Maybe<Scalars['Boolean']>;
    /** A list of labels associated with the object. */
    labels?: Maybe<LabelConnection>;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars['DateTime']>;
    /** `true` if the object is locked */
    locked: Scalars['Boolean'];
    /** Identifies the milestone associated with the issue. */
    milestone?: Maybe<Milestone>;
    /** Identifies the issue number. */
    number: Scalars['Int'];
    /** A list of Users that are participating in the Issue conversation. */
    participants: UserConnection;
    /** List of project cards associated with this issue. */
    projectCards: ProjectCardConnection;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars['DateTime']>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue */
    resourcePath: Scalars['URI'];
    /** Identifies the state of the issue. */
    state: IssueState;
    /**
     * A list of events, comments, commits, etc. associated with the issue.
     * @deprecated `timeline` will be removed Use Issue.timelineItems instead. Removal on 2020-10-01 UTC.
     */
    timeline: IssueTimelineConnection;
    /** A list of events, comments, commits, etc. associated with the issue. */
    timelineItems: IssueTimelineItemsConnection;
    /** Identifies the issue title. */
    title: Scalars['String'];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars['DateTime'];
    /** The HTTP URL for this issue */
    url: Scalars['URI'];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Can user react to this subject */
    viewerCanReact: Scalars['Boolean'];
    /** Check if the viewer is able to change their subscription status for the repository. */
    viewerCanSubscribe: Scalars['Boolean'];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars['Boolean'];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars['Boolean'];
    /** Identifies if the viewer is watching, not watching, or ignoring the subscribable entity. */
    viewerSubscription?: Maybe<SubscriptionState>;
  };
// ...
export enum PullRequestState {
  /** A pull request that is still open. */
  Open = 'OPEN',
  /** A pull request that has been closed without being merged. */
  Closed = 'CLOSED',
  /** A pull request that has been closed by being merged. */
  Merged = 'MERGED',
}
// ...
export type PullRequestInfoFragment = { __typename?: 'PullRequest' } & Pick<
  PullRequest,
  'id' | 'number' | 'title' | 'url'
> & { prState: PullRequest['state'] } & {
    repository: { __typename?: 'Repository' } & RepoInfoFragment;
  };

export type IssueInfoFragment = { __typename?: 'Issue' } & Pick<
  Issue,
  'id' | 'number' | 'title' | 'url'
> & { issueState: Issue['state'] } & {
    repository: { __typename?: 'Repository' } & RepoInfoFragment;
  };

type SearchResult_App_Fragment = { __typename: 'App' };

type SearchResult_Issue_Fragment = { __typename: 'Issue' } & IssueInfoFragment;

type SearchResult_MarketplaceListing_Fragment = {
  __typename: 'MarketplaceListing';
};

type SearchResult_Organization_Fragment = { __typename: 'Organization' };

type SearchResult_PullRequest_Fragment = {
  __typename: 'PullRequest';
} & PullRequestInfoFragment;

type SearchResult_Repository_Fragment = { __typename: 'Repository' };

type SearchResult_User_Fragment = { __typename: 'User' };

export type SearchResultFragment =
  | SearchResult_App_Fragment
  | SearchResult_Issue_Fragment
  | SearchResult_MarketplaceListing_Fragment
  | SearchResult_Organization_Fragment
  | SearchResult_PullRequest_Fragment
  | SearchResult_Repository_Fragment
  | SearchResult_User_Fragment;